Добавила компонент user-info
