+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,18 +1,13 @@
 import Logo from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 import FavouriteOfferCard from '../../components/favourite-offers-card/favourite-offers-card';
-import { Link } from 'react-router-dom';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { AuthorizationStatus } from '../../const';
-import { logoutAction } from '../../store/api-actions';
+import {useAppSelector } from '../../hooks';
+import UserInfo from '../../components/user-info/user-info';
 
 function FavoritesScreen () : JSX.Element {
   const offersFromStore = useAppSelector((state) => state.offers);
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const filterOffers = offersFromStore.filter((offer) => offer.isFavorite === true);
   const filterOffer = filterOffers.map((offer) => <FavouriteOfferCard offer={offer} key = {offer.id}/>);
-
-  const dispatch = useAppDispatch();
 
   return (
     <div className="page">
@@ -27,32 +22,7 @@ function FavoritesScreen () : JSX.Element {
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link
-                    className="header__nav-link header__nav-link--profile"
-                    to={'/favorites'}
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    {authorizationStatus === AuthorizationStatus.Auth &&
-                    <>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                      <span className="header__favorite-count">{filterOffer.length}</span>
-                    </>}
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  {authorizationStatus === AuthorizationStatus.Auth ?
-                    <Link className="header__nav-link" to={'/'} onClick={(evt) => {
-                      evt.preventDefault();
-                      dispatch(logoutAction());
-                    }}
-                    >
-                      <span className="header__signout">Sign out</span>
-                    </Link> :
-                    <Link className="header__nav-link" to={'/login'}>
-                      <span className="header__signout">Sign in</span>
-                    </Link>}
-                </li>
+                <UserInfo/>
               </ul>
             </nav>
           </div>
