+++ project/src/pages/main-empty-screen/main-empty-screen.tsx
@@ -1,5 +1,6 @@
 import Logo from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
+import UserInfo from '../../components/user-info/user-info';
 
 function MainEmptyScreen () : JSX.Element {
   return (
@@ -15,23 +16,7 @@ function MainEmptyScreen () : JSX.Element {
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="/"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                  Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="/">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
+                <UserInfo/>
               </ul>
             </nav>
           </div>
