+++ project/src/pages/not-found-screen/not-found-screen.tsx
@@ -1,13 +1,10 @@
 import { Link } from 'react-router-dom';
 import Logo from '../../components/logo/logo';
-import {AppRoute, AuthorizationStatus} from '../../const';
+import {AppRoute} from '../../const';
 import {Helmet} from 'react-helmet-async';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { logoutAction } from '../../store/api-actions';
+import UserInfo from '../../components/user-info/user-info';
 
 function NotFoundScreen () : JSX.Element {
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
-  const dispatch = useAppDispatch();
 
   return (
     <div className="page page--gray">
@@ -22,30 +19,7 @@ function NotFoundScreen () : JSX.Element {
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={'/favorites'}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    {authorizationStatus === AuthorizationStatus.Auth &&
-                    <>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                      <span className="header__favorite-count">3</span>
-                    </>}
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  {authorizationStatus === AuthorizationStatus.Auth ?
-                    <Link className="header__nav-link" to={'/'} onClick={(evt) => {
-                      evt.preventDefault();
-                      dispatch(logoutAction());
-                    }}
-                    >
-                      <span className="header__signout">Sign out</span>
-                    </Link> :
-                    <Link className="header__nav-link" to={'/login'}>
-                      <span className="header__signout">Sign in</span>
-                    </Link>}
-                </li>
+                <UserInfo/>
               </ul>
             </nav>
           </div>
