+++ project/src/pages/property-screen/property-screen.tsx
@@ -1,29 +1,25 @@
 import Logo from '../../components/logo/logo';
 import {Helmet} from 'react-helmet-async';
 import CommentForm from '../../components/comment-form/comment-form';
-import { Link, useParams } from 'react-router-dom';
+import {useParams } from 'react-router-dom';
 import {reviews} from '../../mocks/reviews';
 import {nearOffers} from '../../mocks/near-offer';
 import ReviewList from '../../components/review-list/review-list';
 import OfferList from '../../components/offer-list/offer-list';
 import Map from '../../components/map/map';
-import {AuthorizationStatus, CITIES} from '../../const';
-import { useAppDispatch, useAppSelector } from '../../hooks';
+import {CITIES} from '../../const';
+import {useAppSelector } from '../../hooks';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
-import { logoutAction } from '../../store/api-actions';
-
+import UserInfo from '../../components/user-info/user-info';
 
 function PropertyScreen () : JSX.Element {
   const offersFromStore = useAppSelector((state) => state.offers);
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const currentCity = useAppSelector((state) => state.city);
   const params = useParams();
   const currentOffer = offersFromStore.find((offer) => params.id === String(offer.id)); //оффер который отображается в property-screen
   const filteredCity = CITIES.filter((city) => city.title === currentCity);
   const filteredNearOffers = nearOffers.filter((nearOffer) => nearOffer.city.name === currentCity);
   const newOffers = currentOffer ? [...filteredNearOffers, currentOffer] : [];
-
-  const dispatch = useAppDispatch();
 
   if (!currentOffer) {
     return <NotFoundScreen/>;
@@ -41,32 +37,7 @@ function PropertyScreen () : JSX.Element {
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link
-                    className="header__nav-link header__nav-link--profile"
-                    to={'/favorites'}
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    {authorizationStatus === AuthorizationStatus.Auth &&
-                    <>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                      <span className="header__favorite-count">3</span>
-                    </>}
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  {authorizationStatus === AuthorizationStatus.Auth ?
-                    <Link className="header__nav-link" to={'/'} onClick={(evt) => {
-                      evt.preventDefault();
-                      dispatch(logoutAction());
-                    }}
-                    >
-                      <span className="header__signout">Sign out</span>
-                    </Link> :
-                    <Link className="header__nav-link" to={'/login'}>
-                      <span className="header__signout">Sign in</span>
-                    </Link>}
-                </li>
+                <UserInfo/>
               </ul>
             </nav>
           </div>
